:construction: Update month conversion to short for chart

+++ packages/frontend/src/utils/reformatChart.js
@@ -51,7 +51,7 @@ export default nps => {
   const convertDate = date => {
     return new Date(date).toLocaleDateString('en-US', {
       year: 'numeric',
-      month: 'long'
+      month: 'short'
     });
   };
